Consolidate the three attendees-list suites into one mirrored suite
The attendees browser was tested twice: the mirrored direct suite at
test/features/admin/attendees-list.test.ts and the two HTTP suites at
test/integration/server/attendees-{list,csv}.test.ts checked the same
rules (auth, the listing filter, sort order, paging, the CSV export),
with separate fixtures that could drift apart.

The three files merge into test/features/admin/attendees-list/ (page,
filters, rows, csv, plus shared helpers), keeping the stronger variant
of each duplicated rule and every unique case. The merged suite drives
the real routes through the session helpers, like the attendee-notes
suite that moved the same way. Closes the "Two suites now cover the
attendees list" TODO entry.
