+++ test/features/admin/attendees-list/csv.test.ts
@@ -1,22 +1,27 @@
+/**
+ * The attendees CSV export: who may download it, which bookings it includes,
+ * and the record it leaves behind.
+ */
+
 import { expect } from "@std/expect";
 import { describe, it as test } from "@std/testing/bdd";
+import { getAllActivityLog } from "#shared/db/activity-log.ts";
+import { ATTENDEES_PAGE_SIZE } from "#shared/db/attendees/queries.ts";
 import { testRequiresAuth } from "#test-utils/assertions.ts";
 import { describeWithEnv } from "#test-utils/db.ts";
+import { seedFillerAttendees } from "#test-utils/db-helpers/attendee-seeding.ts";
 import {
   createMultiBookingAttendee,
   createTestAttendeeDirect,
 } from "#test-utils/db-helpers/attendees.ts";
 import { createTestListing } from "#test-utils/db-helpers/listings.ts";
-import { adminGet } from "#test-utils/session.ts";
+import { adminGet, withTestSession } from "#test-utils/session.ts";
+import { makeListing, seedListingFilterPair } from "./helpers.ts";
 
-const makeListing = (name: string) =>
-  createTestListing({
-    maxAttendees: 100,
-    name,
-    thankYouUrl: "https://example.com",
-  });
+const csvBody = async (query = ""): Promise<string> =>
+  await (await adminGet(`/admin/attendees/csv${query}`)).text();
 
-describeWithEnv("server (admin attendees CSV)", { db: true }, () => {
+describeWithEnv("the attendees CSV export", { db: true }, () => {
   describe("GET /admin/attendees/csv", () => {
     testRequiresAuth("/admin/attendees/csv");
 
@@ -34,15 +39,16 @@ describeWithEnv("server (admin attendees CSV)", { db: true }, () => {
       expect(csv).toContain("Gala Night");
     });
 
+    test("includes every attendee when no listing is chosen", async () => {
+      await seedListingFilterPair();
+      const csv = await csvBody();
+      expect(csv).toContain("AliceOne");
+      expect(csv).toContain("BobTwo");
+    });
+
     test("filters the export to a single listing", async () => {
-      const first = await makeListing("First");
-      const second = await makeListing("Second");
-      await createTestAttendeeDirect(first.id, "AliceOne", "a1@example.com");
-      await createTestAttendeeDirect(second.id, "BobTwo", "b2@example.com");
-      const response = await adminGet(
-        `/admin/attendees/csv?listing=${first.id}`,
-      );
-      const csv = await response.text();
+      const { first } = await seedListingFilterPair();
+      const csv = await csvBody(`?listing=${first.id}`);
       expect(csv).toContain("AliceOne");
       expect(csv).not.toContain("BobTwo");
     });
@@ -55,10 +61,7 @@ describeWithEnv("server (admin attendees CSV)", { db: true }, () => {
         { listingId: other.id },
       ]);
 
-      const response = await adminGet(
-        `/admin/attendees/csv?listing=${wanted.id}`,
-      );
-      const csv = await response.text();
+      const csv = await csvBody(`?listing=${wanted.id}`);
       // One row: the booking on the filtered listing only — the grouped
       // attendees PAGE shows their other listings, but the export must not.
       expect(csv).toContain("DoubleBooker");
@@ -70,12 +73,35 @@ describeWithEnv("server (admin attendees CSV)", { db: true }, () => {
     test("returns just the header when the type filter matches no listings", async () => {
       const listing = await makeListing("Only Standard");
       await createTestAttendeeDirect(listing.id, "Lonely", "l@example.com");
-      const response = await adminGet("/admin/attendees/csv?type=daily");
-      const csv = await response.text();
+      const csv = await csvBody("?type=daily");
       expect(csv).not.toContain("Lonely");
       // No matching listings → no rows, so only the header line is emitted.
       expect(csv.split("\n")).toHaveLength(1);
       expect(csv).toContain("Listing");
+    });
+
+    test("keeps reading pages until it has every booking", async () => {
+      const listing = await createTestListing({
+        maxAttendees: ATTENDEES_PAGE_SIZE * 3,
+      });
+      // One page's worth plus a few, so an export that stopped after the
+      // first page — or walked backwards — would come back short.
+      const total = ATTENDEES_PAGE_SIZE + 5;
+      await seedFillerAttendees(listing.id, total);
+
+      const csv = await csvBody();
+      const rows = csv.trim().split("\n").length - 1;
+      expect(rows).toBe(total);
+    });
+
+    test("notes that the attendee list was exported", async () => {
+      const listing = await makeListing("Gala Night");
+      await createTestAttendeeDirect(listing.id, "Alice", "alice@example.com");
+      await csvBody();
+      const entries = await withTestSession(() => getAllActivityLog(20));
+      expect(entries.map((entry) => entry.message)).toContain(
+        "Attendees CSV exported",
+      );
     });
   });
 });
